Moved AND, OR and NOT filters out of Filter class in support of serialization
