+++ gcm4/src/main/java/plugins/partitions/testsupport/attributes/support/AttributeFilter.java
@@ -7,9 +7,9 @@ import java.util.Set;
 import nucleus.NucleusError;
 import nucleus.SimulationContext;
 import plugins.partitions.support.Equality;
-import plugins.partitions.support.Filter;
 import plugins.partitions.support.FilterSensitivity;
 import plugins.partitions.support.PartitionError;
+import plugins.partitions.support.filters.Filter;
 import plugins.partitions.testsupport.attributes.AttributesDataManager;
 import plugins.partitions.testsupport.attributes.events.AttributeUpdateEvent;
 import plugins.people.support.PersonId;
