+++ gcm4/src/test/java/plugins/groups/support/AT_GroupMemberFilter.java
@@ -21,8 +21,8 @@ import plugins.groups.events.GroupMembershipRemovalEvent;
 import plugins.groups.testsupport.GroupsTestPluginFactory;
 import plugins.groups.testsupport.GroupsTestPluginFactory.Factory;
 import plugins.groups.testsupport.TestGroupTypeId;
-import plugins.partitions.support.Filter;
 import plugins.partitions.support.FilterSensitivity;
+import plugins.partitions.support.filters.Filter;
 import plugins.people.datamanagers.PeopleDataManager;
 import plugins.people.support.PersonError;
 import plugins.people.support.PersonId;
