+++ gcm4/src/test/java/plugins/groups/support/AT_GroupsForPersonAndGroupTypeFilter.java
@@ -21,9 +21,9 @@ import plugins.groups.testsupport.GroupsTestPluginFactory;
 import plugins.groups.testsupport.GroupsTestPluginFactory.Factory;
 import plugins.groups.testsupport.TestGroupTypeId;
 import plugins.partitions.support.Equality;
-import plugins.partitions.support.Filter;
 import plugins.partitions.support.FilterSensitivity;
 import plugins.partitions.support.PartitionError;
+import plugins.partitions.support.filters.Filter;
 import plugins.people.datamanagers.PeopleDataManager;
 import plugins.people.support.PersonError;
 import plugins.people.support.PersonId;
