+++ gcm4/src/test/java/plugins/partitions/support/AT_PopulationPartitionImpl.java
@@ -23,6 +23,7 @@ import nucleus.ActorContext;
 import nucleus.Event;
 import nucleus.SimulationContext;
 import nucleus.testsupport.testplugin.TestSimulation;
+import plugins.partitions.support.filters.Filter;
 import plugins.partitions.testsupport.FunctionalAttributeLabeler;
 import plugins.partitions.testsupport.PartitionsTestPluginFactory;
 import plugins.partitions.testsupport.PartitionsTestPluginFactory.Factory;
