+++ gcm4/src/test/java/plugins/partitions/support/filters/AT_Filter.java
@@ -1,4 +1,4 @@
-package plugins.partitions.support;
+package plugins.partitions.support.filters;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -14,6 +14,8 @@ import org.junit.jupiter.api.Test;
 import nucleus.Event;
 import nucleus.SimulationContext;
 import nucleus.testsupport.testplugin.TestSimulation;
+import plugins.partitions.support.FilterSensitivity;
+import plugins.partitions.support.PartitionError;
 import plugins.partitions.testsupport.PartitionsTestPluginFactory;
 import plugins.partitions.testsupport.PartitionsTestPluginFactory.Factory;
 import plugins.people.datamanagers.PeopleDataManager;
@@ -70,22 +72,22 @@ public class AT_Filter {
 			assertEquals(100, peopleDataManager.getPopulationCount());
 
 			// create the filters
-			Filter filter = Filter.allPeople().and(Filter.allPeople());
+			Filter filter = new TrueFilter().and(new TrueFilter());
 			for (PersonId personId : peopleDataManager.getPeople()) {
 				assertTrue(filter.evaluate(c, personId));
 			}
 
-			filter = Filter.allPeople().and(Filter.noPeople());
+			filter = new TrueFilter().and(new FalseFilter());
 			for (PersonId personId : peopleDataManager.getPeople()) {
 				assertFalse(filter.evaluate(c, personId));
 			}
 
-			filter = Filter.noPeople().and(Filter.allPeople());
+			filter = new FalseFilter().and(new TrueFilter());
 			for (PersonId personId : peopleDataManager.getPeople()) {
 				assertFalse(filter.evaluate(c, personId));
 			}
 
-			filter = Filter.noPeople().and(Filter.noPeople());
+			filter = new FalseFilter().and(new FalseFilter());
 			for (PersonId personId : peopleDataManager.getPeople()) {
 				assertFalse(filter.evaluate(c, personId));
 			}
@@ -107,7 +109,7 @@ public class AT_Filter {
 			// precondition tests
 
 			// if the filter is null
-			ContractException contractException = assertThrows(ContractException.class, () -> Filter.allPeople().and(null));
+			ContractException contractException = assertThrows(ContractException.class, () -> new TrueFilter().and(null));
 			assertEquals(PartitionError.NULL_FILTER, contractException.getErrorType());
 		});
 		TestSimulation.builder().addPlugins(factory.getPlugins()).build().execute();
@@ -131,22 +133,22 @@ public class AT_Filter {
 			assertEquals(100, peopleDataManager.getPopulationCount());
 
 			// create the filters
-			Filter filter = Filter.allPeople().or(Filter.allPeople());
+			Filter filter = new TrueFilter().or(new TrueFilter());
 			for (PersonId personId : peopleDataManager.getPeople()) {
 				assertTrue(filter.evaluate(c, personId));
 			}
 
-			filter = Filter.allPeople().or(Filter.noPeople());
+			filter = new TrueFilter().or(new FalseFilter());
 			for (PersonId personId : peopleDataManager.getPeople()) {
 				assertTrue(filter.evaluate(c, personId));
 			}
 
-			filter = Filter.noPeople().or(Filter.allPeople());
+			filter = new FalseFilter().or(new TrueFilter());
 			for (PersonId personId : peopleDataManager.getPeople()) {
 				assertTrue(filter.evaluate(c, personId));
 			}
 
-			filter = Filter.noPeople().or(Filter.noPeople());
+			filter = new FalseFilter().or(new FalseFilter());
 			for (PersonId personId : peopleDataManager.getPeople()) {
 				assertFalse(filter.evaluate(c, personId));
 			}
@@ -169,7 +171,7 @@ public class AT_Filter {
 
 			// if the filter is null
 
-			ContractException contractException = assertThrows(ContractException.class, () -> Filter.allPeople().or(null));
+			ContractException contractException = assertThrows(ContractException.class, () -> new TrueFilter().or(null));
 			assertEquals(PartitionError.NULL_FILTER, contractException.getErrorType());
 
 		});
@@ -178,11 +180,11 @@ public class AT_Filter {
 	}
 
 	/**
-	 * Tests {@link Filter#negate()}
+	 * Tests {@link Filter#not()}
 	 */
 	@Test
-	@UnitTestMethod(target = Filter.class, name = "negate", args = {})
-	public void testNegate() {
+	@UnitTestMethod(target = Filter.class, name = "not", args = {})
+	public void testNot() {
 		Factory factory = PartitionsTestPluginFactory.factory(100, 4038710674336002107L, (c) -> {
 			PeopleDataManager peopleDataManager = c.getDataManager(PeopleDataManager.class);
 			/*
@@ -191,13 +193,13 @@ public class AT_Filter {
 			 */
 			assertEquals(100, peopleDataManager.getPopulationCount());
 
-			Filter filter = Filter.allPeople().negate();
+			Filter filter = new TrueFilter().not();
 
 			for (PersonId personId : peopleDataManager.getPeople()) {
 				assertFalse(filter.evaluate(c, personId));
 			}
 
-			filter = Filter.noPeople().negate();
+			filter = new FalseFilter().not();
 			for (PersonId personId : peopleDataManager.getPeople()) {
 				assertTrue(filter.evaluate(c, personId));
 			}
@@ -205,49 +207,5 @@ public class AT_Filter {
 			assertEquals(filter.getFilterSensitivities().size(), 0);
 		});
 		TestSimulation.builder().addPlugins(factory.getPlugins()).build().execute();
-	}
-
-	/**
-	 * Tests {@link Filter#allPeople()}
-	 */
-	@Test
-	@UnitTestMethod(target = Filter.class, name = "allPeople", args = {})
-	public void testAllPeople() {
-		Factory factory = PartitionsTestPluginFactory.factory(30, 847391904888351863L, (c) -> {
-			PeopleDataManager peopleDataManager = c.getDataManager(PeopleDataManager.class);
-			// show that the test is valid
-			assertTrue(peopleDataManager.getPopulationCount() > 0);
-
-			final Filter filter = Filter.allPeople();
-
-			for (PersonId personId : peopleDataManager.getPeople()) {
-				assertTrue(filter.evaluate(c, personId));
-			}
-			assertEquals(filter.getFilterSensitivities().size(), 0);
-		});
-		TestSimulation.builder().addPlugins(factory.getPlugins()).build().execute();
-	}
-
-	/**
-	 * Tests {@link Filter#noPeople()}
-	 */
-	@Test
-	@UnitTestMethod(target = Filter.class, name = "noPeople", args = {})
-	public void testNoPeople() {
-		Factory factory = PartitionsTestPluginFactory.factory(100, 6400633994679307999L, (c) -> {
-			PeopleDataManager peopleDataManager = c.getDataManager(PeopleDataManager.class);
-			assertEquals(100, peopleDataManager.getPopulationCount());
-
-			final Filter filter = Filter.noPeople();
-
-			for (PersonId personId : peopleDataManager.getPeople()) {
-				assertFalse(filter.evaluate(c, personId));
-			}
-
-			assertEquals(filter.getFilterSensitivities().size(), 0);
-		});
-		
-		TestSimulation.builder().addPlugins(factory.getPlugins()).build().execute();
-
 	}
 }
