+++ gcm4/src/test/java/plugins/regions/support/AT_RegionFilter.java
@@ -12,8 +12,8 @@ import org.junit.jupiter.api.Test;
 import nucleus.NucleusError;
 import nucleus.SimulationContext;
 import nucleus.testsupport.testplugin.TestSimulation;
-import plugins.partitions.support.Filter;
 import plugins.partitions.support.FilterSensitivity;
+import plugins.partitions.support.filters.Filter;
 import plugins.people.datamanagers.PeopleDataManager;
 import plugins.people.support.PersonId;
 import plugins.regions.datamanagers.RegionsDataManager;
