+++ gcm4/src/test/java/plugins/resources/support/AT_ResourceFilter.java
@@ -13,9 +13,9 @@ import nucleus.NucleusError;
 import nucleus.SimulationContext;
 import nucleus.testsupport.testplugin.TestSimulation;
 import plugins.partitions.support.Equality;
-import plugins.partitions.support.Filter;
 import plugins.partitions.support.FilterSensitivity;
 import plugins.partitions.support.PartitionError;
+import plugins.partitions.support.filters.Filter;
 import plugins.people.datamanagers.PeopleDataManager;
 import plugins.people.support.PersonId;
 import plugins.regions.datamanagers.RegionsDataManager;
